% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eggplots-package.R
\docType{package}
\name{eggplots-package}
\alias{eggplots}
\alias{eggplots-package}
\title{eggplots: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/johnaclouse/eggplots}
  \item Report bugs at \url{https://github.com/johnaclouse/eggplots/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
