% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgraphr.R
\name{hgraphr}
\alias{hgraphr}
\title{hgraphr}
\usage{
hgraphr(df)
}
\arguments{
\item{df}{data frame containing values to be plotted}
}
\description{
A plot for showing high dimensional attributes with indication of percentil ranges.
}
\examples{
df = data.frame(
x = factor(state.name[1:18]),
y = runif(18)
)
hgraphr(df)
}
