% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgraphr.R
\name{hgraphr}
\alias{hgraphr}
\title{Title}
\usage{
hgraphr(df)
}
\arguments{
\item{element_blank}{}
}
\value{

}
\description{
Title
}
\examples{
df = data.frame(
x = factor(state.name[1:18]),
y = runif(18)
)
hgraphr(df)
}
